Fixed Rettangoli() and Trapezi() formula, respectively in Trapezi.cpp and Rettangoli.cpp, substitutingsomma*=h; by putting somma+=h*(f(a+i*h))

+++ Header.h
@@ -14,11 +14,11 @@ void printByteSizes ();
 
 void FloatDoubleDefault(int j);
 
-void Rettangoli(double h,double a,double b,int n);
+void Rettangoli(double h,double a,double b,double n);
 //void SommaRet(double h ,double a,int n);
 
-void Trapezi(double h,double a,double b,int n);
+void Trapezi(double h,double a,double b,double n);
 //void SommaTrap(double h, double a, int n);
 
-void Simpson(double h,double a,double b,int n);
+void Simpson(double h,double a,double b,double n);
 //void SommaSimp(double h, double a, int n);
